Allow day toggling from list screen
I want the user to be able to change the days right from the alarm list
screen. This also will implement a loading state for the list screen's
UI so that if there are ever enough alarms to load that it takes time
it'll at least show a nice spinner.